chore (ci): add running cypress tests
