+++ .bazelignore
@@ -1,0 +1,2 @@
+.tmp/
+node_modules/
