+++ BUILD.bazel
@@ -1,5 +1,3 @@
 package(default_visibility = ["//visibility:public"])
 
-exports_files(["tsconfig.json", "cypress.json"])
-
-
+exports_files(["tsconfig.json"])
