--- cypress.json
@@ -1,7 +1,0 @@
-{
-  "baseUrl": "http://0.0.0.0:3335",
-  "chromeWebSecurity": false,
-  "integrationFolder": "./e2e/integration",
-  "supportFile": "./e2e/support/commands.ts",
-  "testFiles": "**/*spec.ts"
-}
