+++ e2e/BUILD.bazel
@@ -1,44 +1,3 @@
 package(default_visibility = ["//visibility:public"])
 
-load("@cypress//:index.bzl", "cypress_web_test")
-load("@npm//@bazel/typescript:index.bzl", "ts_library")
-
-# You must create a cypress plugin in order to boot a server to serve your application. It can be written as a javascript file or in typescript using ts_library or ts_project.
-ts_library(
-    name = "plugins_file",
-    testonly = True,
-    srcs = ["plugin.ts"],
-    tsconfig = "//:tsconfig.json",
-    deps = [
-        "@npm//@types",
-        "@npm//express",
-    ],
-)
-
-# You can write your cypress tests a javascript files or in typescript using ts_library or ts_project.
-ts_library(
-    name = "hello_spec",
-    testonly = True,
-    srcs = ["hello.spec.ts"],
-    tsconfig = "//:tsconfig.json",
-    deps = [
-        "@npm//cypress",
-    ],
-)
-
-cypress_web_test(
-    # The name of your test target
-    name = "test",
-    srcs = [
-        # Load javascript test files directly as sources
-        "world.spec.js",
-        # Load ts_library tests as a target to srcs
-        ":hello_spec",
-    ],
-    # A cypress config file is required
-    config_file = "//:cypress.json",
-    # Any runtime dependencies you need to boot your server or run your tests
-    data = ["@npm//rxjs"],
-    # Your cypress plugin used to configure cypress and boot your server
-    plugins_file = ":plugins_file",
-)
+exports_files(["cypress.json"])
