+++ e2e/cypress.json
@@ -1,0 +1,4 @@
+{
+  "baseUrl": "http://0.0.0.0:3335",
+  "chromeWebSecurity": false
+}
