+++ e2e/integration/project-selection.spec.ts
@@ -1,25 +1,46 @@
 describe('Select Project', () => {
-
     it('show one project after login', () => {
 
-        cy.visit('');
+        cy.visit('/');
 
-        cy.login('anything.user01@example.org', 'test');
+        // login
+        // cy.request('POST', 'http://0.0.0.0:3333/v2/authentication', {email: 'anything.user01@example.org', password: 'test'});
+        cy.get('#dologin').click();
+        cy.get('#user_id').type('anything.user01@example.org');
+        cy.get('#password').type('test');
+        cy.get('#login_button').click();
+        cy.get('#userctrl').should(($userInfo: JQuery<HTMLElement>) => {
+            expect($userInfo.get(0).innerText).to.contain('User : ');
+        });
 
         // "anything"
         cy.get('#projectctrl').should(($project: JQuery<HTMLElement>) => {
             expect($project.get(0).innerText).to.eq('Project : anything');
         })
 
-        cy.logout();
-
+        // logout
+        // cy.request('DELETE','/v2/authentication');
+        cy.get('#dologout').click();
+        cy.get('#logout_button').click();
+        cy.get('#userctrl').should(($userInfo: JQuery<HTMLElement>) => {
+            expect($userInfo.get(0).innerText).to.eq('');
+        });
     });
 
     it('show several projects after login', () => {
 
-        cy.visit('');
+        cy.visit('/');
 
-        cy.login('multi.user@example.com', 'test');
+        // login
+        // cy.request('POST', 'http://0.0.0.0:3333/v2/authentication', {email: 'multi.user@example.com', password: 'test'});
+        cy.get('#dologin').click();
+        cy.get('#user_id').type('multi.user@example.com');
+        cy.get('#password').type('test');
+        cy.get('#login_button').click();
+        cy.get('#userctrl').should(($userInfo: JQuery<HTMLElement>) => {
+            expect($userInfo.get(0).innerText).to.contain('User : ');
+        });
+
 
         // "anything"
         cy.get('#projectctrl').should(($projects: JQuery<HTMLElement>) => {
@@ -33,8 +54,13 @@ describe('Select Project', () => {
 
         })
 
-        cy.logout();
-
+        // logout
+        // cy.request('DELETE','http://0.0.0.0:3333/v2/authentication');
+        cy.get('#dologout').click();
+        cy.get('#logout_button').click();
+        cy.get('#userctrl').should(($userInfo: JQuery<HTMLElement>) => {
+            expect($userInfo.get(0).innerText).to.eq('');
+        });
     });
 
 });
