+++ e2e/support/BUILD.bazel
@@ -1,15 +1,14 @@
 package(default_visibility = ["//visibility:public"])
 
-load("@cypress//:index.bzl", "cypress_web_test")
 load("@npm//@bazel/typescript:index.bzl", "ts_library")
 
-# You must create a cypress plugin in order to boot a server to serve your application. It can be written as a javascript file or in typescript using ts_library or ts_project.
 ts_library(
-    name = "commands",
+    name = "index",
     testonly = True,
-    srcs = ["commands.ts"],
+    srcs = ["index.ts"],
     tsconfig = "//:tsconfig.json",
     deps = [
+        "@npm//@types",
         "@npm//cypress"
     ],
 )
