+++ e2e/support/commands.ts
@@ -24,27 +24,4 @@
 // -- This will overwrite an existing command --
 // Cypress.Commands.overwrite("visit", (originalFn, url, options) => { ... })
 
-// login
-Cypress.Commands.add("login", (name: string, password: string) => {
-    cy.get('#dologin').click();
 
-    cy.get('#user_id').type(name);
-    cy.get('#password').type(password);
-
-    cy.get('#login_button').click();
-
-    cy.get('#userctrl').should(($userInfo: JQuery<HTMLElement>) => {
-        expect($userInfo.get(0).innerText).to.contain('User : ');
-    });
-});
-
-// logout
-Cypress.Commands.add("logout", () => {
-    cy.get('#dologout').click();
-
-    cy.get('#logout_button').click();
-
-    cy.get('#userctrl').should(($userInfo: JQuery<HTMLElement>) => {
-        expect($userInfo.get(0).innerText).to.eq('');
-    });
-});
