+++ e2e/support/index.ts
@@ -16,11 +16,21 @@
  *  You should have received a copy of the GNU Affero General Public
  *  License along with Knora.  If not, see <http://www.gnu.org/licenses/>.
  */
+import * as express from "express";
 
-describe('hello', () => {
-    it('should find hello', () => {
-        cy.visit('http://localhost:3000');
+module.exports = (on: any, config: any) => {
+    const app = express();
 
-        cy.contains('hello');
+    app.use("/", express.static("./app/public"));
+
+    app.get("/", (req:express.Request, res:express.Response) => {
+        res.sendFile("index.html", { root: "./app/public" });
     });
-});
+
+    const port = 3335;
+    app.listen(port);
+
+    config.baseUrl = `http://0.0.0.0:${port}`;
+
+    return config;
+};
